Update local README + demo  (#94)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
- updated the README so that folks can follow these instructions and
test out our worker-controller!
- also updated the skaffold file so that building the controller image
and deploying to a local k8s cluster is pain-free.

## Why?
- to partially solve #88; the first half where I clean up the conceptual
overview (README present in the root of this repo) shall be coming out
soon!

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->
- fixes #88 

3. How was this tested:
- ran all these commands locally and they work

4. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->