Remove infinite version existence check (#98)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
What it says in the title.
Replaced with Version Status == NotRegistered condition in planning
phase.
Also reorganized test helper functions, ensured that Deployments created
by test helpers can be found by the controller, and added a failed and
successful Gate workflow test.

## Why?
Because checking for version existence in execplan.go breaks the
abstraction of only reading server status in gen status. And we have the
information to not do another describe, so we should use it.

Changes to test files are so that helper functions and test functions
are separated and easier to read and work on.

## Checklist
<!--- add/delete as needed --->

1. Closes #57 

2. How was this tested:
New tests specific to Gate workflows ensure that we are not starting
those workflows on a non-existent version, and old tests also show that
SetCurrent and SetRamping calls are not erroring due to nonexistent
version.

3. Any docs updates needed?
no