Refactor integration tests so they can be run one at a time in IDE (#155)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed and why?
Change from map to a slice of objects with names. GoLand will only put
the little green arrow next to individual table tests if Go can know
what the test name is before runtime. So this is necessary to be able to
run or debug these tests one at a time in the IDE.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
Existing tests run as they did before

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->