Ownership docs: Update the docs to reflect the right command. (#161)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
- Update the ownership transfer command to now reflect the right
command. Previously, it was passing the value for the key
`temporal.io/ignore-last-modifier` as the boolean true.
- This was wrong because the controller expects the value to be a
string.

## Why?
- Correctness.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
- Wasn't tbh

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->